GSYE-392: Rename aggregated_ssp_profile to aggregated_ssp
